SBfeatureSearchOnly {SBadapter}	R Documentation

Run SparkBeyond feature enrichment and learning process.

Description

Run SparkBeyond feature enrichment and learning process.

Usage

SBfeatureSearchOnly(sessionName, trainingFilePath, target, weightColumn = NA,
  maxDepth = 2, hints = NA, useGraph = FALSE, maxFeaturesCount = 300,
  runBlocking = TRUE, server_port = 9000)

Arguments

sessionName
String of the session name.

trainingFilePath
String of the path to the file to be trained on.

target
String of the column name of in the training file that conatins the target of the prediction.

weightColumn
Optional. String of the name of of one of the column that indicate a weighting that is assigned to each example. NA by default.

maxDepth
Optional. Integer < 8 which represent the maximun number of transformations allowed during the feature search phase. Increasing this value should be considered with cautious as the feature search phase is exponential. 2 by default.

hints
Optional. A list of strings that reprents a set of hints that will be used to guide the feature search. NA by default.

useGraph
Optional. A boolean indicating whether the knowledge graph should be used. FALSE by default.

maxFeaturesCount
Optional. An integer of how many features should be created by the SB engine. 300 by default.

server_port
Optional. the port to be accessed in the SparkBeyond API server. 9000 by default.

Value

SBmodel object that encapsulate the feature search result.

Examples

#model = SBfeatureSearchOnly("titanic", titanic_train_filename, "survived")

[Package SBadapter version 0.1 Index]