SBlearn {SBadapter}	R Documentation

Run SparkBeyond feature enrichment and learning process.

Description

Run SparkBeyond feature enrichment and learning process.

Usage

SBlearn(sessionName, trainingFilePath, target, testFilePath = NA,
  trainTestSplitRatio = 0.8, weightColumn = NA, maxDepth = 2,
  algorithmsWhiteList = NA, hints = NA, useGraph = FALSE,
  maxFeaturesCount = 300, evaluationMetric = "PREC",
  scoreOnTestSet = FALSE, crossValidation = 5, runBlocking = TRUE,
  server_port = 9000)

Arguments

sessionName
String of the session name.

trainingFilePath
String of the path to the file to be trained on.

target
String of the column name of in the training file that conatins the target of the prediction.

server_port
Optional. the port to be accessed in the SparkBeyond API server. 9000 by default.

testFilePath:
Optional. String of the path to an independent test file to test the prediction on. NA by default.

trainTestSplitRatio:
Optional. Double value in [0,1] to split the train file data in order to keep some data for test. 0.8 by default. Ignored if test filename was provided.

weightColumn:
Optional. String of the name of of one of the column that indicate a weighting that is assigned to each example. NA by default.

maxDepth:
Optional. Integer < 8 which represent the maximun number of transformations allowed during the feature search phase. Increasing this value should be considered with cautious as the feature search phase is exponential. 2 by default.

algorithmsWhiteList:
Optional. A list of strings that represents the set of algorithms to run. NA by default

hints:
Optional. A list of strings that reprents a set of hints that will be used to guide the feature search. NA by default.

useGraph:
Optional. A boolean indicating whether the knowledge graph should be used. FALSE by default.

maxFeaturesCount:
Optional. An integer of how many features should be created by the SB engine. 300 by default.

evaluationMetric:
Optional. A string representing the evaluation metric. Should be either "AUC", "PREC", or "RMSE". "PREC" by default.

scoreOnTestSet:
Optional. A boolean representing whether scoring should be provided for the test set. FALSE by default.

crossValidation:
Optional. Integer value representing how many cross validation splits should be used. 5 by default.

Value

SBmodel object the encapsulate the prediction.

Examples

model = SBlearn("titanic", getTitanicFilename(train = TRUE), "survived", algorithmsWhiteList = list("RRandomForest"))

## [1] "Artifact location was created at: /Users/zinman/sparkBeyond/coderbot/web/target/universal/stage/localdata/artifacts/titanic/95"
## [1] "In modeling process... please notice that this process may take some time...  0"
## [1] "In modeling process... please notice that this process may take some time...  1"
## [1] "In modeling process... please notice that this process may take some time...  2"
## [1] "In modeling process... please notice that this process may take some time...  3"
## [1] "In modeling process... please notice that this process may take some time...  4"

[Package SBadapter version 0.1 Index]