SBmodel-class {SBadapter}	R Documentation

SB object that encapsulates a model result

Description

SB object that encapsulates a model result

Fields

artifactLoc

String location pointing to the model artifact.

Methods

enrich(dataPath, outputPath, featureCount = NA)

Returns a data frame containing the enrichedData. dataPath is the path to the file to be tested. outputPath is the path to write the results of the prediction. featureCount Integer value signaling how many enriched features would be returned. NA by default - marking maximum number possible (based on the number of features requested in modeling).

evaluate()

Returns an evaluation object containing various information on the run including evaluation metric that was used, evaluation score, precision, confusion matrix, number of correct and incorrect instances, AUC information and more.

initialize(artifact_loc, server_port = 9000)

initializes a model using a string provided as loc.

predict(dataPath, outputPath)

Returns prediction on a created model. dataPath is the path to the file to be tested. outputPath is the path to write the results of the prediction.

showConfusionMatrix(normalized = FALSE, showInIDE = TRUE)

Shows a confusion matrix of a model in the IDE viewer on in the web browser.

showExtractors(showInIDE = TRUE)

Shows extractors used by a model in the IDE viewer on in the web browser.

showFeatures(showInIDE = TRUE)

Shows features used by a model in the IDE viewer on in the web browser.

showFields(showInIDE = TRUE)

Shows fields used by a model in the IDE viewer on in the web browser.

showFunctions(showInIDE = TRUE)

Shows functions used by a model in the IDE viewer on in the web browser.

showInputSchema(showInIDE = TRUE)

Shows the input schema a model in the IDE viewer on in the web browser.

showModelComparison(showInIDE = TRUE)

Shows cross validation of various algorithms tested to create a model in the IDE viewer on in the web browser.

showROC()

Shows ROC of the model in the IDE viewer on in the web browser.

showROC_CV()

Shows ROC of cross validation of various algorithms tested to create a model in the IDE viewer on in the web browser.

showReport(report_name = NA, showInIDE = TRUE)

report_name should be one of the following: confusionMatrix, confusionMatrix_normalized, extractor, features, field, function, InputSchema, modelComparison, roc_best, roc_CV

Examples

#model learn
model = SBlearn("titanic", getTitanicFilename(train = TRUE), "survived",algorithmsWhiteList = list("RRandomForest"))

## [1] "Artifact location was created at: /Users/zinman/sparkBeyond/coderbot/web/target/universal/stage/localdata/artifacts/titanic/96"
## [1] "In modeling process... please notice that this process may take some time...  0"
## [1] "In modeling process... please notice that this process may take some time...  1"
## [1] "In modeling process... please notice that this process may take some time...  2"

#model = SBfeatureSearchOnly("titanic", getTitanicFilename(train = TRUE), "survived")
enriched = model$enrich(getTitanicFilename(train = FALSE), paste(getwd(),"titanic_test_enriched.tsv.gz",sep="/"))
colnames(enriched)

##   [1] "pclass"                                                       
##   [2] "name"                                                         
##   [3] "sex"                                                          
##   [4] "age"                                                          
##   [5] "sibsp"                                                        
##   [6] "parch"                                                        
##   [7] "ticket"                                                       
##   [8] "fare"                                                         
##   [9] "cabin"                                                        
##  [10] "embarked"                                                     
##  [11] "ngrams2_name_contains_mr_"                                    
##  [12] "numericSubstring_ticket_120393_25_111538_75"                  
##  [13] "fare_length_5_0_1_5"                                          
##  [14] "name_zipWithIndex_getOrElse_0_18"                             
##  [15] "numericSubstring_fare_15_172899999999998"                     
##  [16] "sex_male"                                                     
##  [17] "newCore_datastructures_Histogram_name_getOrElse_e_0_0_1_5"    
##  [18] "prefixes_ticket_Contains_349"                                 
##  [19] "prefixes_ticket_Contains_34"                                  
##  [20] "smallSubsets_fare_ContainsList_4_1_7_"                        
##  [21] "ticket_zipWithIndex_getOrElse_A_2147483647_2"                 
##  [22] "newCore_datastructures_Histogram_name_getOrElse_0_0_4_0"      
##  [23] "ngrams2_name_Contains_"                                       
##  [24] "embarked_C"                                                   
##  [25] "fare_zipWithIndex_getOrElse_5_0_3_5_2_0"                      
##  [26] "cabin_"                                                       
##  [27] "pclass_3"                                                     
##  [28] "fare_zipWithIndex_getOrElse_3_0_4_0_0_5"                      
##  [29] "numericSuffix_fare_4438_0_4416_0"                             
##  [30] "newCore_datastructures_Histogram_name_exists___M_2_0_"        
##  [31] "name_zipWithIndex_getOrElse_A_0_34_75_5_25"                   
##  [32] "name_zipWithIndex_getOrElse_P_0_8_5_6_0"                      
##  [33] "prefixes_fare_Contains_26_2"                                  
##  [34] "smallSubsets_ticket_ContainsList_3_4_6_"                      
##  [35] "name_zipWithIndex_getOrElse_R_0_10_25_4_25"                   
##  [36] "name_zipWithIndex_contains_28_"                               
##  [37] "name_zipWithIndex_getOrElse_o_0_12_75_1_75"                   
##  [38] "name_indicesContains29"                                       
##  [39] "name_zipWithIndex_getOrElse_y_2147483647_8"                   
##  [40] "name_zipWithIndex_getOrElse_p_0_23"                           
##  [41] "name_zipWithIndex_getOrElse_v_0_12_0_5_5"                     
##  [42] "newCore_datastructures_Histogram_fare_getOrElse_4_0_0_1_5"    
##  [43] "ticket_zipWithIndex_contains_8_9_"                            
##  [44] "newCore_datastructur_e_PositiveInfinity_4_5_671004324"        
##  [45] "name_zipWithIndex_contains_11_"                               
##  [46] "ticket_zipWithIndex_contains_1_3_"                            
##  [47] "smallSubsets_fare_ContainsList_0_2_9_"                        
##  [48] "fare_zipWithIndex_getOrElse_6_2147483647_2"                   
##  [49] "fare_zipWithIndex_getOrElse_3_0_2"                            
##  [50] "name_zipWithIndex_getOrElse_F_0_28"                           
##  [51] "name_zipWithIndex_getOrElse_V_2147483647_14"                  
##  [52] "name_zipWithIndex_contains_M_23_"                             
##  [53] "tokenizeByCommonDelimiters_fare_Contains_05"                  
##  [54] "smallSubsets_fare_ContainsList_8_2_9_"                        
##  [55] "fare_zipWithIndex_contains_3_6_"                              
##  [56] "newCore_datastructur_e_PositiveInfinity_2_5_1853096548"       
##  [57] "name_zipWithIndex_getOrElse_u_0_35_0_4_5"                     
##  [58] "ticket_zipWithIndex_contains_9_7_"                            
##  [59] "prefixes_ticket_Contains_117"                                 
##  [60] "name_zipWithIndex_getOrElse_x_0_20"                           
##  [61] "name_zipWithIndex_contains_h_13_"                             
##  [62] "name_zipWithIndex_contains_t_9_"                              
##  [63] "numericSuffix_fare_125_0"                                     
##  [64] "name_zipWithIndex_getOrElse_G_0_5_25_3_25"                    
##  [65] "ticket_zipWithIndex_contains_7_4_"                            
##  [66] "smallSubsets_ticket_ContainsList_7_A_"                        
##  [67] "name_zipWithIndex_contains_r_10_"                             
##  [68] "name_zipWithIndex_contains_k_24_"                             
##  [69] "name_zipWithIndex_contains_n_34_"                             
##  [70] "name_zipWithIndex_contains_31_"                               
##  [71] "ticket_zipWithIndex_getOrElse_7_0_5"                          
##  [72] "name_zipWithIndex_getOrElse_k_0_40_75_4_25"                   
##  [73] "newCore_datastructures_Histogram_name_exists___s_2_0_"        
##  [74] "normalizedNgramsWithoutStopwords_name_ContainsList_mr_joseph_"
##  [75] "smallSubsets_ticket_ContainsList_1_6_"                        
##  [76] "smallSubsets_ticket_ContainsList_3_5_9_"                      
##  [77] "smallSubsets_ticket_ContainsList_6_"                          
##  [78] "numericPrefix_fare_10_5"                                      
##  [79] "age_zipWithIndex_getOrElse_0_2"                               
##  [80] "name_zipWithIndex_getOrElse_E_0_16"                           
##  [81] "newCore_datastructures_Histogram_name_getOrElse_s_0_0_1_5"    
##  [82] "smallSubsets_fare_ContainsList_1_8_6_"                        
##  [83] "smallSubsets_fare_ContainsList_1_8_"                          
##  [84] "fare_length_7"                                                
##  [85] "smallSubsets_cabin_ContainsList_2_E_"                         
##  [86] "name_zipWithIndex_getOrElse_o_2147483647_28"                  
##  [87] "fare_lastOption_Some_7_"                                      
##  [88] "name_indicesContains44"                                       
##  [89] "name_zipWithIndex_getOrElse_K_1_25_0_6_0"                     
##  [90] "name_zipWithIndex_getOrElse_M_0_24"                           
##  [91] "ticket_zipWithIndex_contains_1_0_"                            
##  [92] "fare_zipWithIndex_contains_4_1_"                              
##  [93] "fare_zipWithIndex_getOrElse_7_0_3_0_0_5"                      
##  [94] "name_zipWithIndex_getOrElse_l_0_28_0_1_5"                     
##  [95] "name_zipWithIndex_getOrElse_T_0_24"                           
##  [96] "newCore_datastructures_Histogram_fare_getOrElse_3_0_0_1_5"    
##  [97] "newCore_datastructures_Histogram_name_getOrElse_a_0_0_2_5"    
##  [98] "ticket_length_6"                                              
##  [99] "numericPrefix_fare_6_0"                                       
## [100] "ticket_zipWithIndex_getOrElse_0_0_3_0_0_5"                    
## [101] "name_zipWithIndex_contains_e_42_"                             
## [102] "newCore_datastructur_e_PositiveInfinity_1_5_227445247"        
## [103] "ticket_zipWithIndex_contains_5_9_"                            
## [104] "fare_zipWithIndex_contains_6_1_"                              
## [105] "ngrams2_name_Contains_miss_Anna"                              
## [106] "name_zipWithIndex_contains_e_34_"                             
## [107] "name_zipWithIndex_getOrElse_0_10_0_6_5"                       
## [108] "tokenizeByCommonDelimitersAndNormalize_name_contains_mr_"     
## [109] "ticket_zipWithIndex_getOrElse_7_0_4_0_0_5"                    
## [110] "name_zipWithIndex_contains_M_22_"                             
## [111] "tokenizeByCommonDelimiters_fare_Contains_8"                   
## [112] "fare_zipWithIndex_contains_1_5_"                              
## [113] "name_zipWithIndex_getOrElse_J_0_15_5_5_0"                     
## [114] "name_zipWithIndex_getOrElse_z_0_35"                           
## [115] "age_sorted_47"                                                
## [116] "fare_length_6"                                                
## [117] "name_indicesContains31"                                       
## [118] "name_zipWithIndex_getOrElse_g_2147483647_6"                   
## [119] "pclass_1"                                                     
## [120] "name_zipWithIndex_getOrElse_p_2147483647_7"                   
## [121] "numericSuffix_cabin_20_0"                                     
## [122] "smallSubsets_cabin_ContainsList_5_2_"                         
## [123] "smallSubsets_cabin_ContainsList_1_D_"                         
## [124] "prefixes_cabin_Contains_E1"                                   
## [125] "mode_cabin_Some_9_"                                           
## [126] "embarked_S"                                                   
## [127] "smallSubsets_fare_ContainsList_0_5_7_"                        
## [128] "name_zipWithIndex_getOrElse_t_2_19"                           
## [129] "name_zipWithIndex_getOrElse_J_0_26"                           
## [130] "newCore_datastructures_Histogram_name_getOrElse_M_0_0_1_5"    
## [131] "tokenizeByCommonDelimiters_name_Contains_j"                   
## [132] "smallSubsets_ticket_ContainsList_3_5_"                        
## [133] "name_zipWithIndex_contains_e_26_"                             
## [134] "name_zipWithIndex_contains_i_7_"                              
## [135] "ticket_zipWithIndex_getOrElse_5_0_9_25_0_75"                  
## [136] "newCore_datastructures_Histogram_name_exists___6_0_"          
## [137] "name_zipWithIndex_contains_S_13_"                             
## [138] "prefixes_ticket_Contains_267"                                 
## [139] "name_zipWithIndex_contains_45_"                               
## [140] "name_zipWithIndex_getOrElse_B_2147483647_46"                  
## [141] "name_zipWithIndex_getOrElse_c_2147483647_4"                   
## [142] "name_zipWithIndex_getOrElse_B_0_8"                            
## [143] "name_length_43"                                               
## [144] "prefixes_ticket_Contains_PC1747"                              
## [145] "name_zipWithIndex_getOrElse_w_2147483647_18_5_1_0"            
## [146] "name_zipWithIndex_getOrElse_n_0_19"                           
## [147] "name_zipWithIndex_contains_g_24_"                             
## [148] "name_zipWithIndex_getOrElse_R_0_15"                           
## [149] "ngrams2_ticket_Contains_s_o"                                  
## [150] "name_zipWithIndex_contains_o_26_"                             
## [151] "name_zipWithIndex_getOrElse_c_0_13_0_4_5"                     
## [152] "name_zipWithIndex_contains_J_26_"                             
## [153] "numericSubstring_ticket_336058_0"                             
## [154] "name_zipWithIndex_getOrElse_s_2147483647_15"                  
## [155] "prefixes_name_Contains_Cald"                                  
## [156] "name_length_20"                                               
## [157] "tokenizeByCommonDelimiters_name_Contains_kate"                
## [158] "numericPrefix_fare_57_0"                                      
## [159] "name_zipWithIndex_contains_i_25_"                             
## [160] "name_zipWithIndex_getOrElse_0_12"                             
## [161] "newCore_datastructures_Histogram_name_exists___i_5_0_"        
## [162] "tokenizeByCommonDelimiters_fare_Contains_65"                  
## [163] "age_sorted_35"                                                
## [164] "name_zipWithIndex_contains_g_8_"                              
## [165] "name_zipWithIndex_getOrElse_b_0_31"                           
## [166] "smallSubsets_ticket_ContainsList_"                            
## [167] "name_zipWithIndex_contains_S_19_"                             
## [168] "name_zipWithIndex_contains_i_9_"                              
## [169] "name_indices_contains_22_"                                    
## [170] "smallSubsets_fare_ContainsList_3_4_8_"                        
## [171] "ticket_zipWithIndex_getOrElse_9_2147483647_2"                 
## [172] "prefixes_name_Contains_Mad"                                   
## [173] "fare_tailContains_3_"                                         
## [174] "newCore_datastructures_Histogram_name_getOrElse_y_0_0_1_5"    
## [175] "name_zipWithIndex_contains_m_5_"                              
## [176] "ticket_zipWithIndex_getOrElse_8_0_10"                         
## [177] "name_zipWithIndex_contains_f_7_"                              
## [178] "smallSubsets_ticket_ContainsList_4_2_"                        
## [179] "ticket_zipWithIndex_getOrElse_1_2147483647_4"                 
## [180] "numericPrefix_age_64_0"                                       
## [181] "name_zipWithIndex_getOrElse_D_0_11_0_1_5"                     
## [182] "name_zipWithIndex_getOrElse_i_0_10_0_1_5"                     
## [183] "name_zipWithIndex_getOrElse_s_0_11_0_3_5"                     
## [184] "newCore_datastructures_Histogram_name_getOrElse_f_0_0_1_5"    
## [185] "fare_zipWithIndex_contains_9_5_"                              
## [186] "newCore_datastructur_ouble_PositiveInfinity_1948126896"       
## [187] "numericSuffix_fare_9_0"                                       
## [188] "smallSubsets_fare_ContainsList_3_1_"                          
## [189] "numericPrefix_age_4_5_2_0"                                    
## [190] "newCore_datastructures_Histogram_cabin_exists___3_2_0_"       
## [191] "ngrams2_name_Contains_miss_Margaret"                          
## [192] "tokenizeBySpace_name_Contains_mary"                           
## [193] "ngrams2_name_Contains_miss_Elizabeth"                         
## [194] "prefixes_fare_Contains_18_"                                   
## [195] "name_length_33"                                               
## [196] "name_zipWithIndex_contains_A_24_"                             
## [197] "cabin_zipWithIndex_getOrElse_5_1_2"                           
## [198] "newCore_datastructur_e_PositiveInfinity_1_5_1989454623"       
## [199] "numericPrefix_age_0_0"                                        
## [200] "newCore_datastructures_Histogram_fare_getOrElse_1_0_0_2_5"    
## [201] "name_zipWithIndex_contains_A_23_"                             
## [202] "name_zipWithIndex_getOrElse_I_2147483647_28_25_6_25"          
## [203] "name_indices_contains_24_"                                    
## [204] "ticket_zipWithIndex_contains_0_8_"                            
## [205] "tokenizeByCommonDelimitersAndNormalize_name_Contains_daniel"  
## [206] "name_zipWithIndex_contains_c_4_"                              
## [207] "ticket_zipWithIndex_contains_4_8_"                            
## [208] "prefixes_name_Contains_Pa"                                    
## [209] "name_zipWithIndex_getOrElse_b_2147483647_17"                  
## [210] "prefixes_fare_Contains_14_4"                                  
## [211] "tokenizeByCommonDelimiters_name_Contains_johan"               
## [212] "ngrams2_name_Contains_elizabeth"                              
## [213] "tokenizeByCommonDelimiters_name_Contains_ada"                 
## [214] "name_indicesContains41"                                       
## [215] "ticket_zipWithIndex_contains_9_5_"                            
## [216] "name_indicesContains30"                                       
## [217] "prefixes_name_Contains_Bou"                                   
## [218] "tokenizeByCommonDelimiters_name_Contains_nils"                
## [219] "prefixes_name_Contains_Sund"                                  
## [220] "smallSubsets_ticket_ContainsList_4_8_6_"                      
## [221] "ticket_zipWithIndex_getOrElse_C_0_1_5_1_0"                    
## [222] "tokenizeByCommonDelimiters_name_Contains_david"               
## [223] "fare_zipWithIndex_getOrElse_2_0_2_0_0_5"                      
## [224] "smallSubsets_fare_ContainsList_3_8_2_"                        
## [225] "name_zipWithIndex_getOrElse_f_2147483647_20"                  
## [226] "prefixes_name_Contains_Beck"                                  
## [227] "prefixes_cabin_Contains_C10"                                  
## [228] "numericPrefix_fare_76_0"                                      
## [229] "smallSubsets_cabin_ContainsList_4_5_"                         
## [230] "prefixes_ticket_Contains_1175"                                
## [231] "prefixes_cabin_Contains_C5"                                   
## [232] "smallSubsets_cabin_ContainsList_6_A_"                         
## [233] "name_zipWithIndex_contains_p_25_"                             
## [234] "smallSubsets_cabin_ContainsList_3_9_"                         
## [235] "name_zipWithIndex_getOrElse_r_2_11_5_4_0"                     
## [236] "name_zipWithIndex_contains_l_41_"                             
## [237] "name_zipWithIndex_contains_o_43_"                             
## [238] "smallSubsets_ticket_ContainsList_1_7_"                        
## [239] "ticket_zipWithIndex_contains_2_5_"                            
## [240] "name_zipWithIndex_contains_m_8_"                              
## [241] "prefixes_ticket_Contains_174"                                 
## [242] "newCore_datastructur_ouble_PositiveInfinity_1386585326"       
## [243] "smallSubsets_ticket_ContainsList_5_P_"                        
## [244] "fare_lastOption_Some_8_"                                      
## [245] "prefixes_name_Contains_Ol"                                    
## [246] "ticket_zipWithIndex_getOrElse_2_0_3_0_2_5"                    
## [247] "smallSubsets_fare_ContainsList_3_2_"                          
## [248] "prefixes_name_Contains_Lan"                                   
## [249] "newCore_datastructur_e_PositiveInfinity_3_0_701785833"        
## [250] "age_zipWithIndex_getOrElse_3_3_1"                             
## [251] "newCore_datastructures_Histogram_name_exists___r_2_0_"        
## [252] "name_zipWithIndex_contains_s_11_"                             
## [253] "newCore_datastructures_Histogram_name_exists___i_4_0_"        
## [254] "fare_sorted_12357"                                            
## [255] "name_zipWithIndex_contains_g_17_"                             
## [256] "ticket_min_3_"                                                
## [257] "newCore_datastructur_e_PositiveInfinity_4_5_1544657881"       
## [258] "name_zipWithIndex_getOrElse_k_0_24"                           
## [259] "smallSubsets_ticket_ContainsList_3_C_6_"                      
## [260] "ticket_zipWithIndex_contains_7_0_"                            
## [261] "name_zipWithIndex_contains_t_25_"                             
## [262] "smallSubsets_ticket_ContainsList_1_S_7_"                      
## [263] "prefixes_ticket_Contains_230"                                 
## [264] "mode_ticket_Some_4_"                                          
## [265] "tokenizeByCommonDelimiters_name_Contains_katherine"           
## [266] "smallSubsets_cabin_ContainsList_D_7_"                         
## [267] "tokenizeByCommonDelimiters_name_Contains_florence"            
## [268] "tokenizeByCommonDelimiters_name_Contains_jane"                
## [269] "tokenizeByCommonDelimiters_name_Contains_marie"               
## [270] "tokenizeByCommonDelimiters_name_Contains_bertha"              
## [271] "tokenizeByCommonDelimiters_name_Contains_emily"               
## [272] "prefixes_cabin_Contains_B7"                                   
## [273] "name_zipWithIndex_contains_t_36_"                             
## [274] "name_zipWithIndex_getOrElse_G_0_28"                           
## [275] "smallSubsets_ticket_ContainsList_3_0_C_"                      
## [276] "prefixes_ticket_Contains_PC174"                               
## [277] "prefixes_name_Contains_Bi"                                    
## [278] "smallSubsets_fare_ContainsList_4_1_9_"                        
## [279] "smallSubsets_ticket_ContainsList_0_9_"                        
## [280] "ngrams2_name_Contains_alfred_"                                
## [281] "mode_ticket_Some_"                                            
## [282] "smallSubsets_fare_ContainsList_2_6_"                          
## [283] "name_zipWithIndex_contains_t_21_"                             
## [284] "name_zipWithIndex_getOrElse_e_2147483647_4_0_0_5"             
## [285] "name_zipWithIndex_contains_u_26_"                             
## [286] "newCore_datastructures_Histogram_ticket_getOrElse_2_0_0_2_5"  
## [287] "smallSubsets_ticket_ContainsList_4_8_9_"                      
## [288] "tokenizeByCommonDelimiters_name_Contains_martha"              
## [289] "name_zipWithIndex_getOrElse_e_0_33"                           
## [290] "name_zipWithIndex_contains_h_38_"                             
## [291] "tokenizeBySpace_name_Contains_mary.1"                         
## [292] "name_zipWithIndex_getOrElse_S_1_32_25_1_25"                   
## [293] "name_zipWithIndex_contains_F_15_"                             
## [294] "smallSubsets_ticket_ContainsList_1_P_"                        
## [295] "prefixes_ticket_Contains_31508"                               
## [296] "prefixes_ticket_Contains_239"                                 
## [297] "ticket_zipWithIndex_getOrElse_0_2147483647_3"                 
## [298] "sibsp_1"                                                      
## [299] "fare_min_5_"                                                  
## [300] "fare_tail_9"                                                  
## [301] "name_zipWithIndex_contains_s_13_"                             
## [302] "ticket_zipWithIndex_contains_4_1_"                            
## [303] "prefixes_name_Contains_Fra"                                   
## [304] "newCore_datastructures_Histogram_name_getOrElse_0_0_4_5"      
## [305] "name_zipWithIndex_getOrElse_u_0_31"                           
## [306] "smallSubsets_ticket_ContainsList_3_4_8_"                      
## [307] "name_zipWithIndex_contains_44_"                               
## [308] "name_zipWithIndex_getOrElse_d_0_43"                           
## [309] "name_zipWithIndex_contains_a_39_"                             
## [310] "name_zipWithIndex_contains_e_45_"

predicted = model$predict(getTitanicFilename(train = FALSE), paste(getwd(),"titanic_test_predicted.tsv.gz",sep="/"))
colnames(predicted)

##  [1] "pclass"             "name"               "sex"               
##  [4] "age"                "sibsp"              "parch"             
##  [7] "ticket"             "fare"               "cabin"             
## [10] "embarked"           "survived_predicted" "X_0_probability"   
## [13] "X_1_probability"

predicted[1:5,c("survived_predicted", "X_0_probability", "X_1_probability")]

##   survived_predicted X_0_probability X_1_probability
## 1                  0           0.928           0.072
## 2                  1           0.480           0.520
## 3                  0           0.798           0.202
## 4                  0           0.954           0.046
## 5                  1           0.330           0.670

eval = model$evaluate()

## === Detailed Accuracy By Class ===
##                                                                                      
## Class           TP Rate   FP Rate   Precision   Recall   F-Measure   ROC Area   Gini 
##                                                                                      
## 0               0.916     0.351     0.79        0.916    0.848       0.875      0.751
## 1               0.649     0.084     0.842       0.649    0.733       0.875      0.751
## Weighted Avg.   0.807     0.242     0.811       0.807    0.801       0.875      0.751
## 

#model$showFeatures()
#model$showConfusionMatrix()

[Package SBadapter version 0.1 Index]