Session-class {SBadapter}	R Documentation

SB object that encapsulates a session

Description

SB object that encapsulates a session

Fields

artifact_loc

String location pointing to the model artifact.

modelBuilt

Indication for whether only a feature search was performed or a full model was created.

Methods

buildNumber()

Returns the build number in which the model was generated.

createPackage(sampleData = NULL)

Create a sharable package for the model. sampleData can be used to a sample data to the package and test it. Only first 20 rows of the sample data will be used.

enrich(data, featureCount = NA, writePredictionColumnsOnly = TRUE)

Returns a data frame containing the enrichedData. data is a dataframe to be enriched. Set featureCount in order to limit the number of returned features. Set writePredictionColumnsOnly to TRUE to return only prediction and probabily columns rather than the entire dataset.

enrich.file(file, featureCount = NA, writePredictionColumnsOnly = TRUE)

Returns a data frame containing the enrichedData. file is a path to the file to be enriched. Set featureCount in order to limit the number of returned features. Set writePredictionColumnsOnly to TRUE to return only prediction and probabily columns rather than the entire dataset.

evaluate()

Returns an evaluation object containing various information on the run including evaluation metric that was used, evaluation score, precision, confusion matrix, number of correct and incorrect instances, AUC information and more.

features()

Returns a dataset with top feature information

initialize(artifact_loc, modelBuilt = FALSE)

initializes a session using a string provided as loc.

liftFromPrediction(predictionResult, overrideDesiredClass = NA, title = NA, percentOfPopulationToPlot = 0.2)

Returns lift from a created model and generates three plots. predictionResult is a dataframe to be analyzed, overrideDesiredClass the class in the label column to check the lift for (e.g. '1'), title optional: a title for the plot. percentOfPopulationToPlot optional: limit the plot to the top percent of the data (x axis).

predict(data, writePredictionColumnsOnly = TRUE)

Returns prediction on a created model. data is a dataframe to be predicted. Set writePredictionColumnsOnly to TRUE to return only prediction and probabily columns rather than the entire dataset.

predict.file(file, writePredictionColumnsOnly = TRUE)

Returns prediction on a created model. file is the path of the file to be predicted. Set writePredictionColumnsOnly to TRUE to return only prediction and probabily columns rather than the entire dataset.

showConfusionMatrix(normalized = FALSE)

Shows a confusion matrix of a model.

showExtractors()

Shows extractors.

showFeatureClusters()

Shows the representative feature clusters pdf report.

showFeaturesTest()

Shows features performance on test.

showFeaturesTrain()

Shows features performance on train.

showFields()

Shows fields.

showFunctions()

Shows functions.

showInputSchema()

Shows the input schema.

showModelComparison()

Shows cross validation of various algorithms tested to create a model.

showROC()

Shows ROC of the model.

showROC_CV()

Shows ROC of cross validation of various algorithms tested to create a model.

showReport(report_name = NA)

report_name should be one of the following: confusionMatrix, confusionMatrix_normalized, extractor, train_features, train_unweighted_features, test_unweighted_features,field, function, InputSchema, modelComparison, roc_best, roc_CV

status()

Checking the status of the session.

waitForProcess()

Blocking the R console until session is finished.

Examples

# Learning example
session = learn("titanic", getData("titanic_train"), target="survived",algorithmsWhiteList = list("xgboostClassifier"),  scoreOnTestSet = TRUE, useCachedFeatures=TRUE)

## [1] "Training on  /tmp/data_in7096302b8447.tsv"
## [1] "Calling: http://127.0.0.1:9000/rapi/learn"
## [1] "Artifact location was created at: /Users/zinman/localdata/artifacts/titanic/16"
## [1] "Detecting types - 1"
## [1] "Detecting types - 2"
## [1] "Detecting types - 3"
## Feature Search training-set: 710.0
## Original histogram (FS):
## +-------+-----------+
## | Label | Frequency |
## +-------+-----------+
## | 0     | 442       |
## | 1     | 268       |
## +-------+-----------+
## Prior entropy: 0.9562317858093743
## 
## Input Schema
## +-------+-----------+---------------+
## | index | name      | type          |
## +-------+-----------+---------------+
## | 0     | *survived | Int           |
## | 1     | pclass    | Int           |
## | 2     | name      | core.nlp.Text |
## | 3     | sex       | String        |
## | 4     | age       | Double        |
## | 5     | sibsp     | Int           |
## | 6     | parch     | Int           |
## | 7     | ticket    | core.nlp.Text |
## | 8     | fare      | Double        |
## | 9     | cabin     | String        |
## | 10    | embarked  | String        |
## +-------+-----------+---------------+
## [1] "13:07:28 INFO  Learner$ - Learning to predict survived as classification"
## [1] "Session in progress:  Creating features - 4"
## [1] "Session in progress:  Creating features - 5"
## [1] "13:07:42 INFO  LocalFeatureGenerator$ - Finished discovering 3 features in pclass"
## [1] "13:07:42 INFO  LocalFeatureGenerator$ - Finished discovering 6 features in parch"
## [1] "13:07:43 INFO  LocalFeatureGenerator$ - Finished discovering 54 features in age"
## [1] "13:07:43 INFO  LocalFeatureGenerator$ - Finished discovering 1 features in sex"
## [1] "Session in progress:  Creating features - 6"
## [1] "Printing top 50 features out of 300"
##     idx
##  1:   0
##  2:   1
##  3:   2
##  4:   3
##  5:   4
##  6:   5
##  7:   6
##  8:   7
##  9:   8
## 10:   9
## 11:  10
## 12:  11
## 13:  12
## 14:  13
## 15:  14
## 16:  15
## 17:  16
## 18:  17
## 19:  18
## 20:  19
## 21:  20
## 22:  21
## 23:  22
## 24:  23
## 25:  24
## 26:  25
## 27:  26
## 28:  27
## 29:  28
## 30:  29
## 31:  30
## 32:  31
## 33:  32
## 34:  33
## 35:  34
## 36:  35
## 37:  36
## 38:  37
## 39:  38
## 40:  39
## 41:  40
## 42:  41
## 43:  42
## 44:  43
## 45:  44
## 46:  45
## 47:  46
## 48:  47
## 49:  48
## 50:  49
##     idx
##                                                                     feature
##  1:                         tokenizeByPunctuation(name) does not contain mr
##  2:                                                             pclass != 3
##  3:                                              fare >= 15.172899999999998
##  4:                                                             sex != male
##  5:              abs(cos(fare) - -0.2994095593470102) < 0.11486522725075635
##  6:                                                             cabin != ""
##  7:                                                         name contains "
##  8:                                                           embarked == C
##  9:                                                             pclass == 1
## 10:               abs(tan(age) - -0.9727759445199381) >= 1.1379791699016033
## 11:                                           upperCaseRatio(name) == 0.125
## 12: abs(upperCaseRatio(ticket) - 0.09196428571428572) < 0.04196428571428571
## 13:                                         cos(fare) < -0.9989278625592983
## 14:                                                       round(fare) == 19
## 15:                              upperCaseRatio(name) < 0.15301003344481606
## 16:                                                numberOfWords(name) == 7
## 17:             abs(sin(fare) - 0.9989864568219151) < 0.0005608210670511315
## 18:                                                numberOfWords(name) >= 6
## 19:                                                    sigmoid(fare) == 1.0
## 20:                                                              sibsp >= 5
## 21:                        normalizedNgrams(name) contains List(mr, joseph)
## 22:                                tokenizeByPunctuation(ticket) contains o
## 23:                                                      floor(fare) < 10.5
## 24:                                 smallSubsets(cabin) contains List(2, E)
## 25:                               tokenizeByPunctuation(name) contains john
## 26:                                                       ceil(fare) == 7.0
## 27:                        normalizedNgrams(name) contains List(miss, anna)
## 28:                stemmedWordsWithoutCommonTerms(name) does not contain mr
## 29:                                                      ceil(fare) == 35.0
## 30:                                            numericSuffix(cabin) == 20.0
## 31:                                 smallSubsets(cabin) contains List(5, 2)
## 32:                                 smallSubsets(cabin) contains List(1, D)
## 33:                                             prefixes(cabin) contains E1
## 34:                                                  mode(cabin) == Some(9)
## 35:                                                           embarked != S
## 36:                                                             age == 47.0
## 37:                                 normalizedNgrams(name) contains List(j)
## 38:                             sentimentDistribution(name) contains (0, 3)
## 39:                                                     name contains david
## 40:                                                       name contains van
## 41:                                                      name contains kate
## 42:                                                     floor(fare) == 57.0
## 43:                                         sin(age) < -0.21049876125431888
## 44:                                                      ceil(fare) == 23.0
## 45:                                                    name contains julian
## 46:                        slidingSubsets(name).toSeq contains (mr, eugene)
## 47:                      stemmedWordsWithoutCommonTerms(name) contains hugh
## 48:                                 normalizedNgrams(name) contains List(e)
## 49:                                                      abs(age - 4.5) < 2
## 50:                   normalizedNgrams(name) contains List(miss, elizabeth)
##                                                                     feature
##             RIG       support
##  1: 0.246524231 299 (42.11%) 
##  2: 0.076231905 315 (44.37%) 
##  3: 0.068179011 333 (46.90%) 
##  4: 0.242667723 251 (35.35%) 
##  5: 0.026757770  72 (10.14%) 
##  6: 0.066467876 163 (22.96%) 
##  7: 0.025171897  44 (6.20%)  
##  8: 0.023689570 139 (19.58%) 
##  9: 0.062929319 172 (24.23%) 
## 10: 0.016217111 336 (47.32%) 
## 11: 0.011422643  12 (1.69%)  
## 12: 0.018753223  26 (3.66%)  
## 13: 0.012511649   6 (0.85%)  
## 14: 0.012511649   6 (0.85%)  
## 15: 0.051727079 345 (48.59%) 
## 16: 0.032384616  29 (4.08%)  
## 17: 0.023655675  31 (4.37%)  
## 18: 0.058777034  82 (11.55%) 
## 19: 0.046715422 163 (22.96%) 
## 20: 0.011189840  11 (1.55%)  
## 21: 0.008115772   8 (1.13%)  
## 22: 0.009393665  16 (2.25%)  
## 23: 0.064727877 299 (42.11%) 
## 24: 0.012511649   6 (0.85%)  
## 25: 0.008115772   8 (1.13%)  
## 26: 0.010163270  10 (1.41%)  
## 27: 0.016722656   8 (1.13%)  
## 28: 0.093387963 197 (27.75%) 
## 29: 0.006075752   6 (0.85%)  
## 30: 0.010413798   5 (0.70%)  
## 31: 0.010413798   5 (0.70%)  
## 32: 0.010413798   5 (0.70%)  
## 33: 0.010413798   5 (0.70%)  
## 34: 0.010413798   5 (0.70%)  
## 35: 0.026285346 208 (29.30%) 
## 36: 0.007094831   7 (0.99%)  
## 37: 0.007094831   7 (0.99%)  
## 38: 0.017400935 279 (39.30%) 
## 39: 0.006075752   6 (0.85%)  
## 40: 0.005058528   5 (0.70%)  
## 41: 0.012511649   6 (0.85%)  
## 42: 0.008321012   4 (0.56%)  
## 43: 0.011542697 248 (34.93%) 
## 44: 0.007401255   7 (0.99%)  
## 45: 0.004150531   2 (0.28%)  
## 46: 0.004150531   2 (0.28%)  
## 47: 0.004043152   4 (0.56%)  
## 48: 0.005662242   6 (0.85%)  
## 49: 0.018603209  16 (2.25%)  
## 50: 0.010413798   5 (0.70%)  
##             RIG       support
## [1] "13:07:43 INFO  LocalFeatureGenerator$ - Finished discovering 4 features in embarked"
## [1] "13:07:43 INFO  LocalFeatureGenerator$ - Finished discovering 7 features in sibsp"
## [1] "13:07:44 INFO  LocalFeatureGenerator$ - Finished discovering 163 features in fare"
## [1] "13:07:44 INFO  LocalFeatureGenerator$ - Finished discovering 132 features in ticket"
## [1] "13:07:44 INFO  LocalFeatureGenerator$ - Finished discovering 199 features in cabin"
## [1] "13:07:45 INFO  LocalFeatureGenerator$ - Finished discovering 490 features in name"
## [1] "13:07:45 INFO  FeatureAggregator - Selecting best features..."
## [1] "Session in progress:  Building model - 7"
## [1] "13:07:59 INFO  FeatureCompiler$ - finished compiling feature extractor"
## [1] "13:07:59 INFO  Enricher$ - Enriching training set with discovered features..."
## [1] "Session in progress:  Building model - 8"
## 
## == Summary ==
##                                                
## Correctly Classified Instances     147   81.22%
## Incorrectly Classified Instances   34    18.78%
## Total Number of Instances          181         
##                                                
## === Detailed Accuracy By Class ===
##                                                                                      
## Class           TP Rate   FP Rate   Precision   Recall   F-Measure   ROC Area   Gini 
##                                                                                      
## 0               0.944     0.378     0.783       0.944    0.856       0.833      0.667
## 1               0.622     0.056     0.885       0.622    0.73        0.833      0.667
## Weighted Avg.   0.812     0.247     0.825       0.812    0.805       0.833      0.667
##                                                                                      
## Root Mean Prior Squared Error:0.0
## Confusion Matrix:
## 
## +-----+----+
## | 101 | 6  |
## | 28  | 46 |
## +-----+----+
## [1] "Done"

#Do Not Run: session = featureSearch("titanic", getData("titanic_train"), target="survived")
enriched = session$enrich(getData("titanic_train"), featureCount = 10)

## [1] "Enriching  /tmp/data_in709664e996a7.tsv"
## [1] "Calling: http://127.0.0.1:9000/rapi/enrich"
## [1] "Done."

colnames(enriched)

##  [1] "X.tokenizeByPunctuation.name..contains.mr"                                                                                                                                                                                                                                                                                                                                                                                                                                                                           
##  [2] "pclass....3"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
##  [3] "fare....15.172899999999998"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          
##  [4] "sex....male"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
##  [5] "abs.cos.fare.....0.2994095593470102....0.11486522725075635"                                                                                                                                                                                                                                                                                                                                                                                                                                                          
##  [6] "cabin...."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           
##  [7] "name.contains....embarked....C.pclass....1.abs.tan.age.....0.9727759445199381.....1.1379791699016033.false.false.false.false.false.false.false.false.false.false.true.true.true.true.false.true.false.true.true.true.true.false.false.true.false.false.false.false.false.true.true.true.true.true.false.true.false.false.true.true..tokenizeByPunctuation.name..contains.mr.pclass....3.fare....15.172899999999998.sex....male.abs.cos.fare.....0.2994095593470102....0.11486522725075635.cabin......name.contains.."
##  [8] "embarked....C"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       
##  [9] "pclass....1"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
## [10] "abs.tan.age.....0.9727759445199381.....1.1379791699016033"

predicted = session$predict(getData("titanic_test"))

## [1] "Predicting  /tmp/data_in709658e8c10b.tsv"
## [1] "Calling: http://127.0.0.1:9000/rapi/predict"
## [1] "Predictions and plots are available at: /Users/zinman/localdata/artifacts/titanic/16/reports/predictions/test/"
## [1] "Done."

colnames(predicted)

## [1] "survived_predicted" "probability_0"      "probability_1"

predicted[1:5,c("survived_predicted", "probability_0", "probability_1")]

##   survived_predicted probability_0 probability_1
## 1                  0     0.9445565     0.0554435
## 2                  1     0.3302991     0.6697009
## 3                  0     0.7451480     0.2548520
## 4                  0     0.8949008     0.1050992
## 5                  1     0.2847269     0.7152731

eval = session$evaluate()

## === Detailed Accuracy By Class ===
##                                                                                      
## Class           TP Rate   FP Rate   Precision   Recall   F-Measure   ROC Area   Gini 
##                                                                                      
## 0               0.944     0.378     0.783       0.944    0.856       0.833      0.667
## 1               0.622     0.056     0.885       0.622    0.73        0.833      0.667
## Weighted Avg.   0.812     0.247     0.825       0.812    0.805       0.833      0.667
## 

#session$showFeatures()
#session$showConfusionMatrix()

[Package SBadapter version 0.1 Index]