featureSearch {SBadapter}	R Documentation

Run SparkBeyond feature enrichment and learning process.

Description

Run SparkBeyond feature enrichment and learning process.

Usage

featureSearch(projectName = "temp", trainData, target, weightColumn = NA,
  maxDepth = 2, functionsWhiteList = NA, functionsBlackList = NA,
  booleanNumericFeatures = TRUE, useGraph = FALSE,
  crossRowFeatureSearch = FALSE, maxFeaturesCount = list(300),
  columnSubsetSize = 1, customColumnSubsets = NA,
  maxFeatureDuration = 100,
  overrideMaxFeatureDurationForExternalData = TRUE,
  useCachedFeatures = FALSE, allocatedMemoryMB = 1000,
  maxCollectionSize = 80000, weightByClass = FALSE,
  produceFeatureClusteringReport = FALSE, fileEncoding = NA,
  runBlocking = TRUE)

Arguments

projectName
Optional string of the session name. Setting a session name is highly recommened. "temp" by default.

target
String of the column name of in the training file that conatins the target of the prediction.

weightColumn
Optional. String of the name of of one of the column that indicate a weighting that is assigned to each example. NA by default.

maxDepth
Optional. Integer < 8 which represent the maximun number of transformations allowed during the feature search phase. Increasing this value should be considered with cautious as the feature search phase is exponential. 2 by default.

useGraph
Optional. A boolean indicating whether the knowledge graph should be used. FALSE by default.

trainData:
A data frame to analyze.

functionsWhiteList:
Optional. A list of strings that represents a set of functions that will be used to guide the feature search. NA by default.

functionsBlackList:
Optional. A list of strings that represents a set of function that will be excluded from the feature search. Can also include function domains including('math','arithmetics', 'collections', 'booleanOperators', 'semantics', 'nlp', 'trigonometry', 'bitwise'). NA by default.

booleanNumericFeatures:
A boolean indicating whether to transform all features to boolean values. TRUE by default.

crossRowFeatureSearch.
A booleean indicating whether to allow creating features using data collected from multiple rows together.FALSE by default.

maxFeaturesCount:
Optional. A list of integers indicating how many features should be created by the SB engine. 300 by default.

columnSubsetSize:
Optional. An integer denoting whether sets of columns should be looked at together. 1 by default.

allocatedMemoryMB:
Optional. Integer value representing how to chunk the memory during feature search . 1000MB by default.

maxCollectionSize:
Optional. Integer value repsenting what is the maximum cardinality allowed for a transformation during feature search. 80K by default.

customColumnSubsets:
Optional. A List of lists containing specific column subsets to examine. NA by default.

maxFeatureDuration:
Optional. A numeric value representing the maximum allowed time a feature may take during search per row in milliseconds. 100 by default.

overrideMaxFeatureDurationForExternalData:
A boolean indicating whether the maxFeatureDuration parameter should not be used for features that use external data. TRUE by default.

useCachedFeatures:
Optional. A boolean indicating whether to use cached features (from previous run). FALSE by default.

weightByClass:
Adds a weight column with values inverse proportional to the frequency of the class. FALSE by default.

produceFeatureClusteringReport:
An indicator to produce feature cluster visualization. FALSE by default.

fileEncoding:
Optional. NA by default. Options are: "ISO-8859-1", "UTF-8", "US-ASCII".

Value

Session object that encapsulate the feature search result.

Examples

#session = featureSearch("titanic", getData("titanic_train"), "survived")

[Package SBadapter version 0.1 Index]