flattenCols {SBadapter}	R Documentation

flattenCols

Description

Take the only the first element from each column defined in cols and set the column type into a vector

Usage

flattenCols(data, cols, keepLast = FALSE)

Arguments

data:
data.table to flatten

cols:
list of columns to flatten

keepLast:
Boolean. Indicates whether to keep the first or last element in the series.

Value

nothing: operates on the input data object

Examples

grouped = groupBy(getData("titanic_train"), by =list("pclass","sex"))
typeof(grouped$age)

## [1] "list"

flattenCols(grouped, "age")
typeof(grouped$age) ## notice that the type of the column changed

## [1] "double"

[Package SBadapter version 0.1 Index]