isLatestRpackage {SBadapter}	R Documentation

A function to check if the current version is the latest one

Description

A function to check if the current version is the latest one

Usage

isLatestRpackage()

[Package SBadapter version 0.1 Index]