runTitanicLearn {SBadapter}	R Documentation

Perform learn on SparkBeyond titanic example.

Description

Perform learn on SparkBeyond titanic example.

Usage

runTitanicLearn(configuration = "1", server_port = 9000)

Arguments

configuration
Configuration ID as string. 1 - use only J48, 2 - use RRandomForest, 3 - use defaultList.

server_port
the port to be accessed in the SparkBeyond API server. 9000 by default.

Value

SBmodel object containing the result of the learning

Examples

model = runTitanicLearn()

## [1] "Artifact location was created at: /Users/zinman/sparkBeyond/coderbot/web/target/universal/stage/localdata/artifacts/titanic/97"
## [1] "In modeling process... please notice that this process may take some time...  0"
## [1] "In modeling process... please notice that this process may take some time...  1"
## [1] "In modeling process... please notice that this process may take some time...  2"

[Package SBadapter version 0.1 Index]