sampleData {SBadapter}	R Documentation

sampleData

Description

Sample data by ratio

Usage

sampleData(data, ratio = 0.8, seed = 1)

Arguments

data
dataframe to be sampled

ratio
the ratio of data to keep

seed
- Optional. A single numeric value representing the seed for the random sampling. It is guaranteed that for the same data and the same seed, the same sampling will be generated.

Value

a sampled dataset

[Package SBadapter version 0.1 Index]