setTimeColumn {SBadapter}	R Documentation

Sugar to trim a list column by a number Sugar to trim a column by another column variable setTimeColumn

Description

Rename a requested column to "SB_times_col". When the exported data is read by the SparkBeyond engine, all other columns that contain lists that are of the same size as of the list in "SB_times_col" will be converted into time series data structures.

Usage

setTimeColumn(data, timeCol)

Arguments

data:
the dataframe / data.table to be modified.

timeCol:
The column name in data to be renamed.

[Package SBadapter version 0.1 Index]