trainTestSplit {SBadapter}	R Documentation

trainTestSplit

Description

Split data to train + test datasets by ratio

Usage

trainTestSplit(data, ratio = 0.8, seed = 1)

Arguments

data
dataframe to be splitted

seed
- Optional. A single numeric value representing the seed for the random sampling. It is guaranteed that for the same data and the same seed, the same sampling will be generated.

Value

a list of two elements with train dataframe as first element, and test dataframe as second element

[Package SBadapter version 0.1 Index]