trimByCol {SBadapter}	R Documentation

Sugar to trim a column by another column variable

Description

Sugar to trim a column by another column variable

Usage

trimByCol(y, n)

[Package SBadapter version 0.1 Index]